+++ NEWS.md
@@ -5,9 +5,10 @@
 
 # Version 2.4.0
 - Functions 'spwb_stress', 'spwb_resistances' and 'spwb_waterUseEfficiency' renamed to 'droughtStress', 'resistances' and 'waterUseEfficiency', respectively, since they can now be applied to the output of several simulation function.
-- Plant biomass balance in growth
-- Bug correction in shrub structural update. Shrub dynamics default set to TRUE
-- IMPORTANT: New species parameter table
+- Plant biomass balance in growth. Structural changes are now daily in growth simulations. 
+- Bug correction in shrub structural update. Shrub dynamics default set to TRUE.
+- DBH/Height plots from growth output.
+- IMPORTANT: New species parameter table.
 
 # Version 2.3.8
 - Summary functions revised, including new function 'summary.fordyn'